refactor: HomeTemplate - 하위 컴포넌트로 관리
