+++ src/components/home/homeTemplate/KidList.tsx
