+++ src/pages/Home/KidHome.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 
 import { useAppDispatch, useAppSelector } from '@store/app/hooks';
 import useAxiosPrivate from '@lib/hooks/auth/useAxiosPrivate';
-import HomeTemplate from '@components/home/HomeTemplate';
+import HomeTemplate from '@components/home/homeTemplate/HomeTemplate';
 import MarginTemplate from '@components/layout/MarginTemplate';
 import { TFetchStatus } from '@lib/types/TFetchStatus';
 import {
